requirements.txt
asgiref==3.11.0
Django==4.2.27
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.4
typing_extensions==4.15.0
whitenoise==6.11.0
requests==2.32.3
weasyprint
matplotlib==3.8.4
dj-database-url==2.1.0
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.5
django-storages==1.14.4
boto3==1.34.162
wagtail==5.2.8
pypdf==4.2.0
pdf2image==1.17.0
Pillow==10.4.0
PyMuPDF==1.24.10
django-axes==6.5.2
sentry-sdk==2.21.0
django-otp==1.5.0
django-two-factor-auth==1.15.2
phonenumbers==8.13.29
django-auditlog==2.3.0
